added validation to prevent deleting event with existing bookings

+++ src/main/java/com/adityaraj/eventbooking/service/EventServiceImpl.java
@@ -1,6 +1,8 @@
 package com.adityaraj.eventbooking.service;
 
+import com.adityaraj.eventbooking.model.Booking;
 import com.adityaraj.eventbooking.model.Event;
+import com.adityaraj.eventbooking.repository.BookingRepository;
 import com.adityaraj.eventbooking.repository.EventRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,11 +29,17 @@ public class EventServiceImpl implements EventService {
         return eventRepository.findById(id).orElseThrow();
     }
 
+    @Autowired
+    BookingRepository bookingRepository;
+
     @Override
     public void deleteById(Long id) {
+        List<Booking> bookings = bookingRepository.findByEventId(id);
+        if(!bookings.isEmpty()) {
+            throw new RuntimeException("Cannot delete event with existing bookings! Cancel all bookings first.");
+        }
         eventRepository.deleteById(id);
     }
-
     @Override
     public List<Event> searchEvents(String keyword) {
         return eventRepository.findByEventNameContainingOrVenueContaining(keyword, keyword);
